Refactor GoVPP usage and generation
